Make SpeedCalculator an interface and ConditionalSpeedCalculator an implementation
